Refactor __init__.py and __main__.py to have smaller functions
